% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridger.R
\name{bridgr}
\alias{bridgr}
\title{Implement bridging correction}
\usage{
bridgr(df, min_grade = NA, max_grade = NA)
}
\arguments{
\item{df}{A dataframe with grading data.}

\item{min_grade}{The minimum possible grade (by default is zero.)}

\item{max_grade}{The maximum possible grade (by default is the
maximum observed grade.)}
}
\value{
\code{df1} A version of the provided grade data.

\code{coregradmeds} The post-processed scores on the scale
of the original grade data.

\code{zmedsrank} The output from \code{bridgr}
}
\description{
Implements a Bayesian version of the Aldrich Mckelvey model to
provided grade data.
}
