% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridgr-package.R
\docType{package}
\name{bridgr-package}
\alias{bridgr}
\alias{bridgr-package}
\title{bridgr: Bridging Data Frequencies for Timely Economic Forecasts}
\description{
To learn more about bridgr, start with the vignettes:
\code{browseVignettes(package = "bridgr")}
}
\details{
\code{bridgr} is under active maintenance. The public interface centered on
\code{\link[=mf_model]{mf_model()}}, \code{\link[=forecast]{forecast()}}, \code{\link[=summary]{summary()}}, and \code{\link[=plot]{plot()}} is considered stable
for bridge-style nowcasting workflows, while future development is expected
to expand diagnostics and mixed-frequency model options without changing the
core target-and-indicator workflow.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/marcburri/bridgr}
  \item \url{https://marcburri.github.io/bridgr/}
  \item Report bugs at \url{https://github.com/marcburri/bridgr/issues}
}

}
\author{
\strong{Maintainer}: Marc Burri \email{marc.burri91@gmail.com} (\href{https://orcid.org/0000-0001-8974-9090}{ORCID}) [copyright holder]

}
\keyword{internal}
