% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridger.R
\name{bridgr.eval.bias}
\alias{bridgr.eval.bias}
\title{Evaluate grading bias}
\usage{
bridgr.eval.bias(df, plot = TRUE, tbl = TRUE)
}
\arguments{
\item{df}{A dataframe with grading data.}

\item{plot}{Whether to plot empirical CDF of grades by grader.}

\item{tbl}{Whether to tabulate best and worst outcomes for bridging observations.}
}
\value{
\code{mae} Vector containing the mean absolute error of grades
and ranks for bridging exams.

\code{rmse} Vector containing the root-mean squared error of
grades and ranks for bridging exams.

\code{mae.bounds} Minimum and maximum values for the MAE of ranks.

\code{fstat} F-statistic and associated p-value for grades.
}
\description{
\code{bridgr.eval.bias} assesses evidence of bias in grading data.
Specifically, this function uses a set of common or bridging exams to
evaluate whether graders systematically differ from each other.
The grading data should should contain the following columns: (a)
one column labeled \code{ta} that lists all graders. Remaining
columns \code{grader1,grader2,..} should be labeled by values
corresponding to the levels of the \code{ta} column.
}
\examples{
\donttest{
sim.midterm = readRDS("simdata_five.rds")
grading.bias.pre = bridgr.eval.bias(df=sim.midterm)
}
}
