% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridger.R
\name{bridgr.eval.bias}
\alias{bridgr.eval.bias}
\title{Evaluate grading bias}
\usage{
bridgr.eval.bias(bridgr.dat, plot = TRUE, tbl = TRUE)
}
\arguments{
\item{bridgr.dat}{A \code{bridgr.data} object.}

\item{plot}{Whether to plot empirical CDF of grades by grader.}

\item{tbl}{Whether to plot tabulate evidence of bias in grading data.}
}
\value{
\code{mae} Vector containing the mean absolute error of grades
and ranks for bridging exams.

\code{rmse} Vector containing the root-mean squared error of
grades and ranks for bridging exams.

\code{bounds} Minimum and maximum values for the MAE and RMSE of ranks.

\code{fstat} F-statistic and associated p-value for grades.
}
\description{
\code{bridgr.eval.bias} assesses evidence of bias in grading data.
Specifically, this function uses a set of common or bridging exams to
evaluate whether graders systematically differ from each other.
}
