% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridger.R
\name{bridgr.eval.post}
\alias{bridgr.eval.post}
\title{Evaluate improvements from bridging process}
\usage{
bridgr.eval.post(bridgr.results, plot = TRUE, tbl = TRUE)
}
\arguments{
\item{bridgr.results}{A \code{bridgr} object.}

\item{plot}{Whether to plot empirical CDF of post-processed (bridged)
grades.}

\item{tbl}{Whether to tabulate evidence of bias in post-processed grades.}
}
\value{
\code{mae} Vector containing the values of the pre-processed
and post-processed mean absolute errors for the grades and ranks

\code{rmse} Vector containing the values of the pre-processed
and post-processed root mean-squared errors for the grades and ranks
}
\description{
\code{bridgr.eval.post} evaluates improvements to grading bias from
the bridging process.
}
\examples{
# Load pre-processed results
data("bridgr.sim.results")
# Evaluate improvements among commonly graded students
grading.bias.post = bridgr.eval.post(bridgr.results = bridgr.sim.results,plot = TRUE, tbl = FALSE)
}
