% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridgr.sim.data.R
\docType{data}
\name{bridgr.sim.data}
\alias{bridgr.sim.data}
\title{A simulated grading dataset.}
\format{
\subsection{\code{bridgr.sim.data}}{

A grading data frame with 142 students and 3 graders:
\describe{
\item{student}{The student associated with the \code{grade} field.}
\item{grader}{The grader or section associated with the \code{grade} field.}
\item{grader.assigned}{The student's assigned grader.}
\item{grade}{The grade to be scaled.}
...
}
}
}
\usage{
bridgr.sim.data
}
\description{
A simulated grading dataset.
}
\keyword{datasets}
