% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridgr.sim.results.R
\docType{data}
\name{bridgr.sim.results}
\alias{bridgr.sim.results}
\title{Example of \code{bridgr::bridgr} output based on the \code{bridgr.sim.data} dataset.}
\format{
\subsection{\code{bridgr.sim.results}}{

\describe{
\item{data}{Processed version of the \code{bridgr.sim.data} dataset.}
\item{student_id}{Student identifier that can be linked with the \code{bridgr.sim.data} dataset.}
\item{zmedsrank}{The post-processed (bridged) student ranks.}
\item{corrected_grades}{The post-processed (bridged) student grades.}
...
}
}
}
\usage{
bridgr.sim.results
}
\description{
Example of \code{bridgr::bridgr} output based on the \code{bridgr.sim.data} dataset.
}
\keyword{datasets}
