% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\docType{data}
\name{forecast.bridge_model}
\alias{forecast.bridge_model}
\title{Implement the \code{forecast} method for \code{bridge_model}}
\format{
An object of class \code{S7_external_generic} of length 4.
}
\usage{
forecast.bridge_model
}
\arguments{
\item{bridge_model}{A \code{bridge} object.}

\item{xreg}{A matrix of external regressors. If not supplied by the user,
the forecast set will be used.}
}
\description{
This method is used to forecast the target variable using the fitted model.
}
\keyword{datasets}
