% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast}
\alias{forecast}
\title{Forecast method for bridge_model
Implement the \code{forecast} method for \code{bridge}}
\usage{
forecast(bridge_model, xreg = NULL, ...)
}
\arguments{
\item{bridge_model}{A \code{bridge} object.}

\item{xreg}{A matrix of external regressors. If not supplied by the user,
the forecast set will be used.}
}
\description{
Forecast method for bridge_model
Implement the \code{forecast} method for \code{bridge}
}
