% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.bridge}
\alias{summary.bridge}
\title{Summarize a Bridge Model}
\usage{
\method{summary}{bridge}(object, ...)
}
\arguments{
\item{object}{A \code{"bridge"} object returned by \code{\link[=bridge]{bridge()}}.}

\item{...}{Unused.}
}
\value{
\code{object}, invisibly.
}
\description{
Summarize a Bridge Model
}
