% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\docType{data}
\name{summary.bridge_model}
\alias{summary.bridge_model}
\title{Summary method for bridge_model}
\format{
An object of class \code{S7_external_generic} of length 4.
}
\usage{
summary.bridge_model
}
\arguments{
\item{bridge_model}{A \code{bridge} object.}
}
\description{
Implement the \code{forecast} method for \code{bridge}
}
\keyword{datasets}
