% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.mf_model}
\alias{summary.mf_model}
\title{Summarize a Mixed-Frequency Model}
\usage{
\method{summary}{mf_model}(object, ...)
}
\arguments{
\item{object}{A \code{"mf_model"} object returned by \code{\link[=mf_model]{mf_model()}}.}

\item{...}{Unused.}
}
\value{
\code{object}, invisibly.
}
\description{
Summarize a Mixed-Frequency Model
}
\examples{
gdp_growth <- suppressMessages(tsbox::ts_na_omit(tsbox::ts_pc(gdp)))
model <- mf_model(
  target = gdp_growth,
  indic = baro,
  indic_predict = "auto.arima",
  indic_aggregators = "mean",
  h = 1
)

summary(model)
}
