% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary}
\alias{summary}
\title{Define the \code{summary} method for \code{bridge_model}}
\arguments{
\item{bridge_model}{A \code{bridge_model} object.}
}
\description{
Define the \code{summary} method for \code{bridge_model}
}
\keyword{This}
\keyword{bridge}
\keyword{internal}
\keyword{is}
\keyword{method}
\keyword{model.}
\keyword{summarizes}
\keyword{the}
