% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridgr-package.R
\docType{package}
\name{bridgr-package}
\alias{bridgr}
\alias{bridgr-package}
\title{bridgr: Bridging Data Frequencies for Timely Economic Forecasts}
\description{
To learn more about bridger, start with the vignettes:
\code{browseVignettes(package = "bridgr")}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/marcburri/bridgr}
  \item \url{https://marcburri.github.io/bridgr/}
  \item Report bugs at \url{https://github.com/marcburri/bridgr/issues}
}

}
\author{
\strong{Maintainer}: Marc Burri \email{marc.burri91@gmail.com} (\href{https://orcid.org/0000-0001-8974-9090}{ORCID}) [copyright holder]

}
\keyword{internal}
