% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridger.R
\name{bridgr.eval.post}
\alias{bridgr.eval.post}
\title{Evaluate improvements from bridging process}
\usage{
bridgr.eval.post(bridgr.results, plot = T, tbl = T)
}
\arguments{
\item{bridgr.results}{The output from \code{bridgr::bridgr}.}

\item{plot}{Whether to plot empirical CDF of post-processed (bridged)
grades.}

\item{tbl}{Whether to tabulate evidence of bias in post-processed grades.}
}
\value{
\code{mae} Vector containing the values of the pre-processed
and post-processed mean absolute errors for the grades and ranks

\code{rmse} Vector containing the values of the pre-processed
and post-processed root mean-squared errors for the grades and ranks
}
\description{
\code{bridgr.eval.post} evaluates improvements to grading bias from
the bridging process.
}
\examples{
\donttest{
bridgr.results = bridgr(df = sim.midterm)
grading.bias.post = bridgr.eval.post(bridgr.results = results)
}
}
