% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridger.R
\name{bridgr}
\alias{bridgr}
\title{Implement bridging correction}
\usage{
bridgr(df, min_grade = NA, max_grade = NA)
}
\arguments{
\item{min_grade}{The minimum possible grade (by default is zero.)}

\item{max_grade}{The maximum possible grade (by default is the
maximum observed grade.)}
}
\value{
\code{df1} A version of the provided grade data.

\code{coregradmeds} The post-processed (bridged) student
grades on the scale of the original grade data.

\code{zmedsrank} The post-processed (bridged) student ranks
}
\description{
Implements a Bayesian version of the Aldrich Mckelvey model to
grade data.
The grading data should contain the following columns: (a)
one column labeled \code{ta} that lists all graders. Remaining
columns \code{grader1,grader2,..} should be labeled by values
corresponding to the levels of the \code{ta} column.
}
\examples{
\donttest{
sim.midterm = readRDS("simdata_five.rds")
grading.bias.pre = bridgr.eval.bias(df=sim.midterm)
}
}
