% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast.bridge}
\alias{forecast.bridge}
\title{Forecasting from a \code{bridge_model} class}
\usage{
\method{forecast}{bridge}(object, xreg = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{bridge_model} object obtained from \code{\link[=bridge]{bridge()}}.}

\item{xreg}{A \code{\link[tsbox:ts_boxable]{tsbox::ts_boxable()}} series of external regressors. If not
supplied by the user, the forecast set calculated by \code{\link[=bridge]{bridge()}}
will be used. This is useful for made up scenarios.}

\item{...}{Additional arguments to be passed to the forecast function.}
}
\description{
This function is used to forecast the target variable using the fitted \code{bridge_model}.
}
