% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast}
\alias{forecast}
\title{Define the \code{forecast} method for \code{bridge_model} class}
\arguments{
\item{bridge_model}{A \code{bridge_model} object.}

\item{xreg}{A matrix of external regressors. If not supplied by the user,
the forecast set will be used. This is useful for made up scenarios.}
}
\description{
Define the \code{forecast} method for \code{bridge_model} class
}
\keyword{This}
\keyword{fitted}
\keyword{forecast}
\keyword{internal}
\keyword{is}
\keyword{method}
\keyword{model.}
\keyword{target}
\keyword{the}
\keyword{to}
\keyword{used}
\keyword{using}
\keyword{variable}
