% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.bridge}
\alias{summary.bridge}
\title{Summarize a \code{bridge_model} class object}
\usage{
\method{summary}{bridge}(object, ...)
}
\arguments{
\item{object}{An object of class \code{bridge_model} obtained from \code{\link[=bridge]{bridge()}}.}

\item{...}{Additional arguments to be passed to the summary function.}
}
\description{
This method is summarizes the bridge model.
}
\keyword{internal}
