% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary}
\alias{summary}
\title{Summary method for bridge_model}
\usage{
summary(bridge_model, ...)
}
\arguments{
\item{bridge_model}{A \code{bridge} object.}
}
\description{
Implement the \code{forecast} method for \code{bridge}
}
